requirements.txt
openai
gtts
RPi.GPIO

